Initial commit for Minefork.
At this point I consider Minefork to be a distinct game from Dreambuilder or MTG,
playable and essentially feature-complete at a basic level (though there are still
features that I want to add), and so I am creating a repository that only contains
Minefork and nothing else.

I have history on a lot of the modules in a private repository, but that repo
includes a lot of things that are not part of or even really relevant to the
Minefork game.  Hence, this new repository.