Technic Addons:  MV Cement Mixer
This is a wet-mixing device that operates next to flowing water, thus
avoiding the need to constantly manage (non-stackable) buckets of water
when crafting wet cement, taking most of the tedium out of the process of
creating substantial amounts of concrete.

For now, wet cement is the only recipe it supports, but in principle
it could support any recipe that mixes two input ingredents plus water.
Currently, concrete (in its various colors) is the only building material
with such a recipe needed to create it.

There are some food recipes, that I never bothered to excise from
dreambuilder, but there is no practical reason to mass-produce large
quantities of those foods; if somebody wants stacks and stacks of
cure-my-fall-dammage they can trivially farm baked potatoes; and if
somebody wants quality levels of food for combat-related reasons they
can implement what they want in conjunction with a combat-related
modpack.  And if people just enjoy the novelty of cooking exotic
foods, they can do that already as it stands, and arguably the fact
that mass production can't be easily automated will make the foods
in question feel more artisanal and thus more desirable in foodie
culture terms.

The terracotta base recipe might make sense to automate if it were
possible to use terracotta base in any way whatosever, but currently
it is not.  You can hold it in your inventory or store it in chests,
but you can't make any building materials out of it, so there's just
no need to produce large amounts of it.