Brick: six slightly different textures that can all tile together.
The fewer sides of each block show, the more of these textures you
can use in your structure (by using a screwdriver to rotate the
block so that a different face shows).  If only one side is visible
(as with e.g. a typical foundation wall), all six can be used, because
it doesn't matter how the other five faces are rotated.  In a more
typical case, the four side faces can be used, by only rotating
the block horizontally (left click with the screwdriver), because
the top and bottom of the block are covered by other blocks and
so don't have to be kept lined up.  By using different faces for
different nodes, the total effect is a significantly more varied
and less repetitive brick wall appearance.